67. vectorize extraction, float32 normalization, and color_seq concat (Phase A perf)
- Vectorize extract_from_location() with zero-padded NumPy advanced indexing
  (replaces per-spot Python loop; expected 5-20x speedup on extraction)
- Use float32 instead of float64 in min_max_normalize() (50% temporary memory reduction)
- Replace .apply(lambda) with .astype(str).agg() for color_seq concatenation
- Add performance optimization plan

+++ src/python/starfinder/preprocessing/normalization.py
@@ -38,7 +38,7 @@ def min_max_normalize(
 
     result = np.empty_like(volume, dtype=np.uint8)
     for c in range(volume.shape[3]):
-        ch = volume[:, :, :, c].astype(np.float64)
+        ch = volume[:, :, :, c].astype(np.float32)
         lo = ch.min()
         hi = ch.max()
